Update and rename problem_020.java to Problem_020.java
